requirements/base.txt
Django==1.11
djangorestframework==3.6.2
read-env==1.1.0
